Organized components (#10)
