+++ frontend/src/components/Login/Login.lazy.js
@@ -1,0 +1,11 @@
+import React, { lazy, Suspense } from 'react';
+
+const LazyLogin = lazy(() => import('./Login'));
+
+const Login = props => (
+  <Suspense fallback={null}>
+    <LazyLogin {...props} />
+  </Suspense>
+);
+
+export default Login;
